refactored matrix ops tests
